Refactored CTA alert into its own component.
